+++ help.go
@@ -1,13 +1,20 @@
 package baker
 
 import (
-	"bytes"
 	"fmt"
 	"io"
-	"os"
-	"reflect"
 	"strings"
-	"unicode"
+)
+
+// HelpFormat represents the possible formats for baker help.
+type HelpFormat int
+
+const (
+	// HelpFormatRaw is for raw-formatted help.
+	HelpFormatRaw HelpFormat = iota
+
+	// HelpFormatMarkdown is for markdown formatted help.
+	HelpFormatMarkdown
 )
 
 // PrintHelp prints the help message for the given component, identified by its name.
@@ -41,226 +48,65 @@ import (
 //     ----------------------------------------------------------------------------------------------------
 //     Listener           | string             |                            | Host:Port to bind to
 //     ----------------------------------------------------------------------------------------------------
-func PrintHelp(w io.Writer, name string, comp Components) {
+func PrintHelp(w io.Writer, name string, comp Components, format HelpFormat) error {
 	dumpall := name == "*"
+
+	generateHelp := GenerateTextHelp
+	if format == HelpFormatMarkdown {
+		generateHelp = GenerateMarkdownHelp
+	}
 
 	for _, inp := range comp.Inputs {
 		if strings.EqualFold(inp.Name, name) || dumpall {
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, "Input: %s\n", inp.Name)
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, inp.Help)
-			if hasConfig(inp.Config) {
-				fmt.Fprintf(w, "\nKeys available in the [input.config] section:\n\n")
-				dumpConfigHelp(w, inp.Config)
-			} else {
-				fmt.Fprintf(w, "\n(no configuration available)\n\n")
+			if err := generateHelp(w, inp); err != nil {
+				return fmt.Errorf("can't print help for %q input: %v", inp.Name, err)
 			}
-			fmt.Fprintln(w)
-			fmt.Fprintln(w)
 			if !dumpall {
-				return
+				return nil
 			}
 		}
 	}
 
 	for _, fil := range comp.Filters {
 		if strings.EqualFold(fil.Name, name) || dumpall {
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, "Filter: %s\n", fil.Name)
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, fil.Help)
-			if hasConfig(fil.Config) {
-				fmt.Fprintf(w, "\nKeys available in the [filter.config] section:\n\n")
-				dumpConfigHelp(w, fil.Config)
-			} else {
-				fmt.Fprintf(w, "\n(no configuration available)\n\n")
+			if err := generateHelp(w, fil); err != nil {
+				return fmt.Errorf("can't print help for %q filter: %v", fil.Name, err)
 			}
-			fmt.Fprintln(w)
-			fmt.Fprintln(w)
 			if !dumpall {
-				return
+				return nil
 			}
 		}
 	}
 
 	for _, out := range comp.Outputs {
 		if strings.EqualFold(out.Name, name) || dumpall {
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, "Output: %s\n", out.Name)
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, out.Help)
-			if hasConfig(out.Config) {
-				fmt.Fprintf(w, "\nKeys available in the [output.config] section:\n\n")
-				dumpConfigHelp(w, out.Config)
-			} else {
-				fmt.Fprintf(w, "\n(no configuration available)\n\n")
-			}
-			fmt.Fprintln(w)
-			fmt.Fprintln(w)
-			if !dumpall {
-				return
+			if strings.EqualFold(out.Name, name) || dumpall {
+				if err := generateHelp(w, out); err != nil {
+					return fmt.Errorf("can't print help for %q output: %v", out.Name, err)
+				}
+				if !dumpall {
+					return nil
+				}
 			}
 		}
 	}
 
 	for _, upl := range comp.Uploads {
 		if strings.EqualFold(upl.Name, name) || dumpall {
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, "Upload: %s\n", upl.Name)
-			fmt.Fprintf(w, "=============================================\n")
-			fmt.Fprintf(w, upl.Help)
-			if hasConfig(upl.Config) {
-				fmt.Fprintf(w, "\nKeys available in the [upload.config] section:\n\n")
-				dumpConfigHelp(w, upl.Config)
-			} else {
-				fmt.Fprintf(w, "\n(no configuration available)\n\n")
-			}
-			fmt.Fprintln(w)
-			fmt.Fprintln(w)
-			if !dumpall {
-				return
+			if strings.EqualFold(upl.Name, name) || dumpall {
+				if err := generateHelp(w, upl); err != nil {
+					return fmt.Errorf("can't print help for %q upload: %v", upl.Name, err)
+				}
+				if !dumpall {
+					return nil
+				}
 			}
 		}
 	}
 
 	if !dumpall {
-		fmt.Fprintf(os.Stderr, "Component not found: %s\n", name)
-	}
-}
-
-func dumpConfigHelp(w io.Writer, cfg interface{}) {
-	const sfmt = "%-18s | %-18s | %-18s | %-8s | "
-	const sep = "----------------------------------------------------------------------------------------------------"
-
-	hpad := fmt.Sprintf(sfmt, "", "", "", "")
-	fmt.Fprintf(w, sfmt, "Name", "Type", "Default", "Required")
-	fmt.Fprintf(w, "Help\n%s\n", sep)
-
-	tf := reflect.TypeOf(cfg).Elem()
-	for i := 0; i < tf.NumField(); i++ {
-		field := tf.Field(i)
-
-		// skip unexported fields
-		if field.PkgPath != "" && !field.Anonymous {
-			continue
-		}
-
-		var typ string
-		switch field.Type.Kind() {
-		case reflect.Int:
-			typ = "int"
-		case reflect.String:
-			typ = "string"
-		case reflect.Slice:
-			switch field.Type.Elem().Kind() {
-			case reflect.String:
-				typ = "array of strings"
-			case reflect.Int:
-				typ = "array of ints"
-			default:
-				panic(field.Type.Elem())
-			}
-		case reflect.Int64:
-			if field.Type.Name() == "Duration" {
-				typ = "duration"
-			} else {
-				typ = "int"
-			}
-		case reflect.Bool:
-			typ = "bool"
-		default:
-			panic(field.Type.Name())
-		}
-
-		help := field.Tag.Get("help")
-		def := field.Tag.Get("default")
-		req := field.Tag.Get("required")
-		if req == "true" {
-			req = "yes"
-		} else {
-			req = "no"
-		}
-
-		fmt.Fprintf(w, sfmt, field.Name, typ, def, req)
-		helpLines := strings.Split(wrapString(help, 60), "\n")
-		if len(helpLines) > 0 {
-			fmt.Fprint(w, helpLines[0], "\n")
-			for _, h := range helpLines[1:] {
-				fmt.Fprint(w, hpad, "  ", h, "\n")
-			}
-		} else {
-			fmt.Fprint(w, "\n")
-		}
+		return fmt.Errorf("component not found: %s", name)
 	}
 
-	fmt.Fprint(w, sep, "\n")
-}
-
-// wrapString wraps the given string within lim width in characters.
-//
-// Source: https://github.com/mitchellh/go-wordwrap
-// Wrapping is currently naive and only happens at white-space. A future
-// version of the library will implement smarter wrapping. This means that
-// pathological cases can dramatically reach past the limit, such as a very
-// long word.
-func wrapString(s string, lim uint) string {
-	// Initialize a buffer with a slightly larger size to account for breaks
-	init := make([]byte, 0, len(s))
-	buf := bytes.NewBuffer(init)
-
-	var current uint
-	var wordBuf, spaceBuf bytes.Buffer
-
-	for _, char := range s {
-		if char == '\n' {
-			if wordBuf.Len() == 0 {
-				if current+uint(spaceBuf.Len()) > lim {
-					current = 0
-				} else {
-					current += uint(spaceBuf.Len())
-					spaceBuf.WriteTo(buf)
-				}
-				spaceBuf.Reset()
-			} else {
-				current += uint(spaceBuf.Len() + wordBuf.Len())
-				spaceBuf.WriteTo(buf)
-				spaceBuf.Reset()
-				wordBuf.WriteTo(buf)
-				wordBuf.Reset()
-			}
-			buf.WriteRune(char)
-			current = 0
-		} else if unicode.IsSpace(char) {
-			if spaceBuf.Len() == 0 || wordBuf.Len() > 0 {
-				current += uint(spaceBuf.Len() + wordBuf.Len())
-				spaceBuf.WriteTo(buf)
-				spaceBuf.Reset()
-				wordBuf.WriteTo(buf)
-				wordBuf.Reset()
-			}
-
-			spaceBuf.WriteRune(char)
-		} else {
-
-			wordBuf.WriteRune(char)
-
-			if current+uint(spaceBuf.Len()+wordBuf.Len()) > lim && uint(wordBuf.Len()) < lim {
-				buf.WriteRune('\n')
-				current = 0
-				spaceBuf.Reset()
-			}
-		}
-	}
-
-	if wordBuf.Len() == 0 {
-		if current+uint(spaceBuf.Len()) <= lim {
-			spaceBuf.WriteTo(buf)
-		}
-	} else {
-		spaceBuf.WriteTo(buf)
-		wordBuf.WriteTo(buf)
-	}
-
-	return buf.String()
+	return nil
 }